+++ src/spot.cpp
@@ -1,7 +1,7 @@
 #include <iostream>
 #include "spot.h"
 
-spot::spot(float X,float Y) : x(X),y(Y) {}
+spot::spot(double X,double Y) : x(X),y(Y) {}
 
 void spot::Print_me(){
   std::cout<<" x : "<<x<<" y : "<<y<<std::endl;
